coreHA | sperate leaderelect function to windows / unix
